Issue #1375: Split build.py into several modules.
