+++ PyInstaller/building/api.py
@@ -1,0 +1,1026 @@
+#-----------------------------------------------------------------------------
+# Copyright (c) 2005-2015, PyInstaller Development Team.
+#
+# Distributed under the terms of the GNU General Public License with exception
+# for distributing bootloader.
+#
+# The full license is in the file COPYING.txt, distributed with this software.
+#-----------------------------------------------------------------------------
+
+
+"""
+This module contains classes that are available for the .spec files.
+
+Spec file is generated by PyInstaller. The generated code from .spec file
+is a way how PyInstaller does the dependency analysis and creates executable.
+"""
+import os
+import shutil
+import tempfile
+
+from PyInstaller import is_win, is_darwin, HOMEPATH, PLATFORM
+from PyInstaller.archive.writers import ZlibArchiveWriter, CArchiveWriter
+from PyInstaller.building.utils import _check_guts_eq, _check_guts_toc_mtime, _check_guts_toc, add_suffix_to_extensions, \
+    checkCache, _check_path_overlap, _rmtree
+from PyInstaller.compat import is_cygwin
+from PyInstaller.depend import bindepend
+from PyInstaller.depend.analysis import get_bootstrap_modules
+from PyInstaller.depend.utils import is_path_to_egg
+from PyInstaller.building.datastruct import TOC, Target, logger
+from PyInstaller.utils import misc
+from PyInstaller.utils.misc import save_py_data_struct, load_py_data_struct
+from .. import log as logging
+
+logger = logging.getLogger(__name__)
+
+if is_win:
+    from PyInstaller.utils.win32 import winmanifest, icon, versioninfo, winresource
+
+
+class PYZ(Target):
+    """
+    Creates a ZlibArchive that contains all pure Python modules.
+    """
+    typ = 'PYZ'
+
+    def __init__(self, toc_dict, name=None, cipher=None):
+        """
+        toc_dict
+            toc_dict['toc']
+                A TOC (Table of Contents), normally an Analysis.pure['toc']?
+            toc_dict['code']
+                A dict of module code objects from ModuleGraph.
+        name
+                A filename for the .pyz. Normally not needed, as the generated
+                name will do fine.
+        cipher
+                The block cipher that will be used to encrypt Python bytecode.
+        """
+
+        from ..config import CONF
+        Target.__init__(self)
+        # TODO remove this attribute, PYZ items are compressed by default.
+        self.compression_level = 0
+        self.toc = toc_dict['toc']
+        # Use code objects directly from ModuleGraph to speed up PyInstaller.
+        self.code_dict = toc_dict['code']
+        self.name = name
+        if name is None:
+            self.name = self.out[:-3] + 'pyz'
+        # PyInstaller bootstrapping modules.
+        self.dependencies = get_bootstrap_modules()
+        # Bundle the crypto key.
+        self.cipher = cipher
+        if cipher:
+            key_file = ('pyimod00_crypto_key',
+                         os.path.join(CONF['workpath'], 'pyimod00_crypto_key.pyc'),
+                         'PYMODULE')
+            # Insert the key as the first module in the list. The key module contains
+            # just variables and does not depend on other modules.
+            self.dependencies.insert(0, key_file)
+        # Compile the top-level modules so that they end up in the CArchive and can be
+        # imported by the bootstrap script.
+        self.dependencies = misc.compile_py_files(self.dependencies, CONF['workpath'])
+        self.__postinit__()
+
+    GUTS = (('name', _check_guts_eq),
+            ('compression_level', _check_guts_eq),
+            ('toc', _check_guts_toc),  # todo: pyc=1
+            )
+
+    def check_guts(self, last_build):
+        if not os.path.exists(self.name):
+            logger.info("Rebuilding %s because %s is missing",
+                        self.outnm, os.path.basename(self.name))
+            return True
+
+        data = Target.get_guts(self, last_build)
+        if not data:
+            return True
+        return False
+
+    def assemble(self):
+        logger.info("Building PYZ (ZlibArchive) %s", os.path.basename(self.out))
+        pyz = ZlibArchiveWriter(code_dict=self.code_dict, cipher=self.cipher)
+        # Do not bundle PyInstaller bootstrap modules into PYZ archive.
+        toc = self.toc - self.dependencies
+        pyz.build(self.name, toc)
+        # FIXME compression level was dropped - remove it from the save_py_data_struct
+        save_py_data_struct(self.out, (self.name, self.compression_level, self.toc))
+        return 1
+
+
+class PKG(Target):
+    """
+    Creates a CArchive. CArchive is the data structure that is embedded
+    into the executable. This data structure allows to include various
+    read-only data in a sigle-file deployment.
+    """
+    typ = 'PKG'
+    xformdict = {'PYMODULE': 'm',
+                 'PYSOURCE': 's',
+                 'EXTENSION': 'b',
+                 'PYZ': 'z',
+                 'PKG': 'a',
+                 'DATA': 'x',
+                 'BINARY': 'b',
+                 'ZIPFILE': 'Z',
+                 'EXECUTABLE': 'b',
+                 'DEPENDENCY': 'd'}
+
+    def __init__(self, toc, name=None, cdict=None, exclude_binaries=0,
+                 strip_binaries=False, upx_binaries=False):
+        """
+        toc
+                A TOC (Table of Contents)
+        name
+                An optional filename for the PKG.
+        cdict
+                Dictionary that specifies compression by typecode. For Example,
+                PYZ is left uncompressed so that it can be accessed inside the
+                PKG. The default uses sensible values. If zlib is not available,
+                no compression is used.
+        exclude_binaries
+                If True, EXTENSIONs and BINARYs will be left out of the PKG,
+                and forwarded to its container (usually a COLLECT).
+        strip_binaries
+                If True, use 'strip' command to reduce the size of binary files.
+        upx_binaries
+        """
+        Target.__init__(self)
+        self.toc = toc
+        self.cdict = cdict
+        self.name = name
+        self.exclude_binaries = exclude_binaries
+        self.strip_binaries = strip_binaries
+        self.upx_binaries = upx_binaries
+        if name is None:
+            self.name = self.out[:-3] + 'pkg'
+        # This dict tells PyInstaller what items embedded in the executable should
+        # be compressed.
+        if self.cdict is None:
+            self.cdict = {'EXTENSION': COMPRESSED,
+                          'DATA': COMPRESSED,
+                          'BINARY': COMPRESSED,
+                          'EXECUTABLE': COMPRESSED,
+                          'PYSOURCE': COMPRESSED,
+                          'PYMODULE': COMPRESSED,
+                          # Do not compress PYZ as a whole. Single modules are
+                          # compressed when creating PYZ archive.
+                          'PYZ': UNCOMPRESSED}
+        self.__postinit__()
+
+    GUTS = (('name', _check_guts_eq),
+            ('cdict', _check_guts_eq),
+            ('toc', _check_guts_toc_mtime),
+            ('exclude_binaries', _check_guts_eq),
+            ('strip_binaries', _check_guts_eq),
+            ('upx_binaries', _check_guts_eq),
+            )
+
+    def check_guts(self, last_build):
+        if not os.path.exists(self.name):
+            logger.info("Rebuilding %s because %s is missing",
+                        self.outnm, os.path.basename(self.name))
+            return 1
+
+        data = Target.get_guts(self, last_build)
+        if not data:
+            return True
+        # todo: toc equal
+        return False
+
+    def assemble(self):
+        logger.info("Building PKG (CArchive) %s", os.path.basename(self.name))
+        trash = []
+        mytoc = []
+        seenInms = {}
+        seenFnms = {}
+        toc = add_suffix_to_extensions(self.toc)
+        # 'inm'  - relative filename inside a CArchive
+        # 'fnm'  - absolute filename as it is on the file system.
+        for inm, fnm, typ in toc:
+            # Ensure filename 'fnm' is not None or empty string. Otherwise
+            # it will fail in case of 'typ' being type OPTION.
+            if fnm and not os.path.isfile(fnm) and is_path_to_egg(fnm):
+                # file is contained within python egg, it is added with the egg
+                continue
+            if typ in ('BINARY', 'EXTENSION', 'DEPENDENCY'):
+                if self.exclude_binaries and typ != 'DEPENDENCY':
+                    self.dependencies.append((inm, fnm, typ))
+                else:
+                    if typ == 'BINARY':
+                        # Avoid importing the same binary extension twice. This might
+                        # happen if they come from different sources (eg. once from
+                        # binary dependence, and once from direct import).
+                        if inm in seenInms:
+                            logger.warn("Two binaries added with the same internal "
+                                        "name. %s was placed at %s previously. "
+                                        "Skipping %s." %
+                                        (seenInms[inm], inm, fnm))
+                            continue
+
+                        # Warn if the same binary extension was included
+                        # with multiple internal names
+                        if fnm in seenFnms:
+                            logger.warn("One binary added with two internal "
+                                        "names. %s was placed at %s previously." %
+                                        (fnm, seenFnms[fnm]))
+                    seenInms[inm] = fnm
+                    seenFnms[fnm] = inm
+
+                    fnm = checkCache(fnm, strip=self.strip_binaries,
+                                     upx=(self.upx_binaries and (is_win or is_cygwin)),
+                                     dist_nm=inm)
+
+                    mytoc.append((inm, fnm, self.cdict.get(typ, 0),
+                                  self.xformdict.get(typ, 'b')))
+            elif typ == 'OPTION':
+                mytoc.append((inm, '', 0, 'o'))
+            else:
+                mytoc.append((inm, fnm, self.cdict.get(typ, 0), self.xformdict.get(typ, 'b')))
+
+        # Bootloader has to know the name of Python library. Pass python libname to CArchive.
+        pylib_name = os.path.basename(bindepend.get_python_library_path())
+        archive = CArchiveWriter(pylib_name=pylib_name)
+
+        archive.build(self.name, mytoc)
+        save_py_data_struct(self.out,
+                   (self.name, self.cdict, self.toc, self.exclude_binaries,
+                    self.strip_binaries, self.upx_binaries))
+        for item in trash:
+            os.remove(item)
+        return 1
+
+
+class EXE(Target):
+    """
+    Creates the final executable of the frozen app.
+    This bundles all necessary files together.
+    """
+    typ = 'EXECUTABLE'
+
+    def __init__(self, *args, **kwargs):
+        """
+        args
+                One or more arguments that are either TOCs Targets.
+        kwargs
+            Possible keywork arguments:
+
+            console
+                On Windows or OSX governs whether to use the console executable
+                or the windowed executable. Always True on Linux/Unix (always
+                console executable - it does not matter there).
+            debug
+                Setting to True gives you progress mesages from the executable
+                (for console=False there will be annoying MessageBoxes on Windows).
+            name
+                The filename for the executable.
+            exclude_binaries
+                Forwarded to the PKG the EXE builds.
+            icon
+                Windows or OSX only. icon='myicon.ico' to use an icon file or
+                icon='notepad.exe,0' to grab an icon resource.
+            version
+                Windows only. version='myversion.txt'. Use grab_version.py to get
+                a version resource from an executable and then edit the output to
+                create your own. (The syntax of version resources is so arcane
+                that I wouldn't attempt to write one from scratch).
+            uac_admin
+                Windows only. Setting to True creates a Manifest with will request
+                elevation upon application restart
+            uac_uiaccess
+                Windows only. Setting to True allows an elevated application to
+                work with Remote Desktop
+        """
+        from ..config import CONF
+        Target.__init__(self)
+
+        # Available options for EXE in .spec files.
+        self.exclude_binaries = kwargs.get('exclude_binaries', False)
+        self.console = kwargs.get('console', True)
+        self.debug = kwargs.get('debug', False)
+        self.name = kwargs.get('name', None)
+        self.icon = kwargs.get('icon', None)
+        self.versrsrc = kwargs.get('version', None)
+        self.manifest = kwargs.get('manifest', None)
+        self.resources = kwargs.get('resources', [])
+        self.strip = kwargs.get('strip', False)
+        # If ``append_pkg`` is false, the archive will not be appended
+        # to the exe, but copied beside it.
+        self.append_pkg = kwargs.get('append_pkg', True)
+
+        # On Windows allows the exe to request admin privileges.
+        self.uac_admin = kwargs.get('uac_admin', False)
+        self.uac_uiaccess = kwargs.get('uac_uiaccess', False)
+
+        if CONF['hasUPX']:
+           self.upx = kwargs.get('upx', False)
+        else:
+           self.upx = False
+
+        # Old .spec format included in 'name' the path where to put created
+        # app. New format includes only exename.
+        #
+        # Ignore fullpath in the 'name' and prepend DISTPATH or WORKPATH.
+        # DISTPATH - onefile
+        # WORKPATH - onedir
+        if self.exclude_binaries:
+            # onedir mode - create executable in WORKPATH.
+            self.name = os.path.join(CONF['workpath'], os.path.basename(self.name))
+        else:
+            # onefile mode - create executable in DISTPATH.
+            self.name = os.path.join(CONF['distpath'], os.path.basename(self.name))
+
+        # Base name of the EXE file without .exe suffix.
+        base_name = os.path.basename(self.name)
+        if is_win or is_cygwin:
+            base_name = os.path.splitext(base_name)[0]
+        self.pkgname = base_name + '.pkg'
+
+        self.toc = TOC()
+
+        ### Pass option -S to Python interpreter to disable the 'site' module.
+        # With this option PyInstaller  won't append user-specific sitedirs to
+        # sys.path.
+        args += ([('S', None, 'OPTION')],)
+
+        for arg in args:
+            if isinstance(arg, TOC):
+                self.toc.extend(arg)
+            elif isinstance(arg, Target):
+                self.toc.append((os.path.basename(arg.name), arg.name, arg.typ))
+                self.toc.extend(arg.dependencies)
+            else:
+                self.toc.extend(arg)
+
+        if is_win:
+            filename = os.path.join(CONF['workpath'], CONF['specnm'] + ".exe.manifest")
+            self.manifest = winmanifest.create_manifest(filename, self.manifest,
+                self.console, self.uac_admin, self.uac_uiaccess)
+            self.toc.append((os.path.basename(self.name) + ".manifest", filename,
+                'BINARY'))
+
+        self.pkg = PKG(self.toc, cdict=kwargs.get('cdict', None),
+                       exclude_binaries=self.exclude_binaries,
+                       strip_binaries=self.strip, upx_binaries=self.upx,
+                       )
+        self.dependencies = self.pkg.dependencies
+        self.__postinit__()
+
+    GUTS = (('name', _check_guts_eq),
+            ('console', _check_guts_eq),
+            ('debug', _check_guts_eq),
+            ('icon', _check_guts_eq),
+            ('versrsrc', _check_guts_eq),
+            ('resources', _check_guts_eq),
+            ('strip', _check_guts_eq),
+            ('upx', _check_guts_eq),
+            ('mtm', None,),  # checked bellow
+            )
+
+    def check_guts(self, last_build):
+        if not os.path.exists(self.name):
+            logger.info("Rebuilding %s because %s missing",
+                        self.outnm, os.path.basename(self.name))
+            return 1
+        if not self.append_pkg and not os.path.exists(self.pkgname):
+            logger.info("Rebuilding because %s missing",
+                        os.path.basename(self.pkgname))
+            return 1
+
+        data = Target.get_guts(self, last_build)
+        if not data:
+            return True
+
+        icon, versrsrc, resources = data[3:6]
+        if (versrsrc or resources) and not is_win:
+            # todo: really ignore :-)
+            logger.warn('ignoring version, manifest and resources, platform not capable')
+        if icon and not (is_win or is_darwin):
+            logger.warn('ignoring icon, platform not capable')
+
+        mtm = data[-1]
+        if mtm != misc.mtime(self.name):
+            logger.info("Rebuilding %s because mtimes don't match", self.outnm)
+            return True
+        if mtm < misc.mtime(self.pkg.out):
+            logger.info("Rebuilding %s because pkg is more recent", self.outnm)
+            return True
+
+        return False
+
+    def _bootloader_file(self, exe):
+        """
+        Pick up the right bootloader file - debug, console, windowed.
+        """
+        # Having console/windowed bootolader makes sense only on Windows and
+        # Mac OS X.
+        if is_win or is_darwin:
+            if not self.console:
+                exe = exe + 'w'
+        # There are two types of bootloaders:
+        # run     - release, no verbose messages in console.
+        # run_d   - contains verbose messages in console.
+        if self.debug:
+            exe = exe + '_d'
+        bootloader_file = os.path.join(HOMEPATH, 'PyInstaller', 'bootloader', PLATFORM, exe)
+        logger.info('Bootloader %s' % bootloader_file)
+        return bootloader_file
+
+    def assemble(self):
+        logger.info("Building EXE from %s", os.path.basename(self.out))
+        trash = []
+        if not os.path.exists(os.path.dirname(self.name)):
+            os.makedirs(os.path.dirname(self.name))
+        outf = open(self.name, 'wb')
+        exe = self._bootloader_file('run')
+        if is_win or is_cygwin:
+            exe = exe + '.exe'
+
+        if not os.path.exists(exe):
+            raise SystemExit(_MISSING_BOOTLOADER_ERRORMSG)
+
+        if is_win and not self.exclude_binaries:
+            # Windows and onefile mode - embed manifest into exe.
+            logger.info('Onefile Mode - Embedding Manifest into EXE file')
+            tmpnm = tempfile.mktemp()
+            shutil.copy2(exe, tmpnm)
+            os.chmod(tmpnm, 0o755)
+
+            # In onefile mode, dependencies in the onefile manifest
+            # refer to files that are about to be unpacked when the exe
+            # is run. The Windows DLL loader doesn't know that and
+            # refuses to run the exe at all. Since the .exe does not in
+            # fact depend on those, and the actual manifest will be used
+            # later when an activation context is created, all
+            # dependencies are removed from the embedded manifest.
+            self.manifest.dependentAssemblies = []
+            self.manifest.update_resources(tmpnm, [1]) # 1 for executable
+            trash.append(tmpnm)
+            exe = tmpnm
+
+        if is_win and (self.icon or self.versrsrc or self.resources):
+            tmpnm = tempfile.mktemp()
+            shutil.copy2(exe, tmpnm)
+            os.chmod(tmpnm, 0o755)
+            if self.icon:
+                icon.CopyIcons(tmpnm, self.icon)
+            if self.versrsrc:
+                versioninfo.SetVersion(tmpnm, self.versrsrc)
+            for res in self.resources:
+                res = res.split(",")
+                for i in range(1, len(res)):
+                    try:
+                        res[i] = int(res[i])
+                    except ValueError:
+                        pass
+                resfile = res[0]
+                restype = resname = reslang = None
+                if len(res) > 1:
+                    restype = res[1]
+                if len(res) > 2:
+                    resname = res[2]
+                if len(res) > 3:
+                    reslang = res[3]
+                try:
+                    winresource.UpdateResourcesFromResFile(tmpnm, resfile,
+                                                        [restype or "*"],
+                                                        [resname or "*"],
+                                                        [reslang or "*"])
+                except winresource.pywintypes.error as exc:
+                    if exc.args[0] != winresource.ERROR_BAD_EXE_FORMAT:
+                        logger.exception(exc)
+                        continue
+                    if not restype or not resname:
+                        logger.error("resource type and/or name not specified")
+                        continue
+                    if "*" in (restype, resname):
+                        logger.error("no wildcards allowed for resource type "
+                                     "and name when source file does not "
+                                     "contain resources")
+                        continue
+                    try:
+                        winresource.UpdateResourcesFromDataFile(tmpnm,
+                                                             resfile,
+                                                             restype,
+                                                             [resname],
+                                                             [reslang or 0])
+                    except winresource.pywintypes.error as exc:
+                        logger.exception(exc)
+            trash.append(tmpnm)
+            exe = tmpnm
+        exe = checkCache(exe, strip=self.strip, upx=self.upx)
+        self.copy(exe, outf)
+        if self.append_pkg:
+            logger.info("Appending archive to EXE %s", self.name)
+            self.copy(self.pkg.name, outf)
+        else:
+            logger.info("Copying archive to %s", self.pkgname)
+            shutil.copy2(self.pkg.name, self.pkgname)
+        outf.close()
+
+        if is_darwin:
+            # Fix Mach-O header for codesigning on OS X.
+            logger.info("Fixing EXE for code signing %s", self.name)
+            from PyInstaller.utils import osxutils
+            osxutils.fix_exe_for_code_signing(self.name)
+            pass
+
+        os.chmod(self.name, 0o755)
+        guts = (self.name, self.console, self.debug, self.icon,
+                self.versrsrc, self.resources, self.strip, self.upx,
+                misc.mtime(self.name))
+        assert len(guts) == len(self.GUTS)
+        save_py_data_struct(self.out, guts)
+        for item in trash:
+            os.remove(item)
+        return 1
+
+    def copy(self, fnm, outf):
+        inf = open(fnm, 'rb')
+        while 1:
+            data = inf.read(64 * 1024)
+            if not data:
+                break
+            outf.write(data)
+
+
+class DLL(EXE):
+    # TODO This feature was dropped and I think it could be removed.
+    """
+    On Windows, this provides support for doing in-process COM servers. It is not
+    generalized. However, embedders can follow the same model to build a special
+    purpose process DLL so the Python support in their app is hidden. You will
+    need to write your own dll.
+    """
+    def assemble(self):
+        logger.info("Building DLL %s", os.path.basename(self.out))
+        outf = open(self.name, 'wb')
+        dll = self._bootloader_file('inprocsrvr') + '.dll'
+        if not os.path.exists(dll):
+            raise SystemExit(_MISSING_BOOTLOADER_ERRORMSG)
+        self.copy(dll, outf)
+        self.copy(self.pkg.name, outf)
+        outf.close()
+        os.chmod(self.name, 0o755)
+        save_py_data_struct(self.out,
+                   (self.name, self.console, self.debug, self.icon,
+                    self.versrsrc, self.manifest, self.resources, self.strip, self.upx, misc.mtime(self.name)))
+        return 1
+
+
+class COLLECT(Target):
+    """
+    In one-dir mode creates the output folder with all necessary files.
+    """
+    def __init__(self, *args, **kws):
+        """
+        args
+                One or more arguments that are either TOCs Targets.
+        kws
+            Possible keywork arguments:
+
+                name
+                    The name of the directory to be built.
+        """
+        from ..config import CONF
+        Target.__init__(self)
+        self.strip_binaries = kws.get('strip', False)
+
+        if CONF['hasUPX']:
+           self.upx_binaries = kws.get('upx', False)
+        else:
+           self.upx_binaries = False
+
+        self.name = kws.get('name')
+        # Old .spec format included in 'name' the path where to collect files
+        # for the created app.
+        # app. New format includes only directory name.
+        #
+        # The 'name' directory is created in DISTPATH and necessary files are
+        # then collected to this directory.
+        self.name = os.path.join(CONF['distpath'], os.path.basename(self.name))
+
+        self.toc = TOC()
+        for arg in args:
+            if isinstance(arg, TOC):
+                self.toc.extend(arg)
+            elif isinstance(arg, Target):
+                self.toc.append((os.path.basename(arg.name), arg.name, arg.typ))
+                if isinstance(arg, EXE):
+                    for tocnm, fnm, typ in arg.toc:
+                        if tocnm == os.path.basename(arg.name) + ".manifest":
+                            self.toc.append((tocnm, fnm, typ))
+                    if not arg.append_pkg:
+                        self.toc.append((os.path.basename(arg.pkgname), arg.pkgname, 'PKG'))
+                self.toc.extend(arg.dependencies)
+            else:
+                self.toc.extend(arg)
+        self.__postinit__()
+
+    GUTS = (('name', _check_guts_eq),
+            ('strip_binaries', _check_guts_eq),
+            ('upx_binaries', _check_guts_eq),
+            ('toc', _check_guts_eq),  # additional check below
+            )
+
+    def check_guts(self, last_build):
+        # COLLECT always needs to be executed, since it will clean the output
+        # directory anyway to make sure there is no existing cruft accumulating
+        return 1
+
+    def assemble(self):
+        if _check_path_overlap(self.name) and os.path.isdir(self.name):
+            _rmtree(self.name)
+        logger.info("Building COLLECT %s", os.path.basename(self.out))
+        os.makedirs(self.name)
+        toc = add_suffix_to_extensions(self.toc)
+        for inm, fnm, typ in toc:
+            if not os.path.exists(fnm) or not os.path.isfile(fnm) and is_path_to_egg(fnm):
+                # file is contained within python egg, it is added with the egg
+                continue
+            if os.pardir in os.path.normpath(inm) or os.path.isabs(inm):
+                raise SystemExit('Security-Alert: try to store file outside '
+                                 'of dist-directory. Aborting. %r' % inm)
+            tofnm = os.path.join(self.name, inm)
+            todir = os.path.dirname(tofnm)
+            if not os.path.exists(todir):
+                os.makedirs(todir)
+            if typ in ('EXTENSION', 'BINARY'):
+                fnm = checkCache(fnm, strip=self.strip_binaries,
+                                 upx=(self.upx_binaries and (is_win or is_cygwin)),
+                                 dist_nm=inm)
+            if typ != 'DEPENDENCY':
+                shutil.copy(fnm, tofnm)
+                try:
+                    shutil.copystat(fnm, tofnm)
+                except OSError:
+                    logger.warn("failed to copy flags of %s", fnm)
+            if typ in ('EXTENSION', 'BINARY'):
+                os.chmod(tofnm, 0o755)
+        save_py_data_struct(self.out,
+                 (self.name, self.strip_binaries, self.upx_binaries, self.toc))
+        return 1
+
+
+class BUNDLE(Target):
+    def __init__(self, *args, **kws):
+        from ..config import CONF
+
+        # BUNDLE only has a sense under Mac OS X, it's a noop on other platforms
+        if not is_darwin:
+            return
+
+        # .icns icon for app bundle.
+        # Use icon supplied by user or just use the default one from PyInstaller.
+        self.icon = kws.get('icon')
+        if not self.icon:
+            self.icon = os.path.join(os.path.dirname(__file__),
+                'bootloader', 'images', 'icon-windowed.icns')
+        # Ensure icon path is absolute.
+        self.icon = os.path.abspath(self.icon)
+
+        Target.__init__(self)
+
+        # .app bundle is created in DISTPATH.
+        self.name = kws.get('name', None)
+        base_name = os.path.basename(self.name)
+        self.name = os.path.join(CONF['distpath'], base_name)
+
+        self.appname = os.path.splitext(base_name)[0]
+        self.version = kws.get("version", "0.0.0")
+        self.toc = TOC()
+        self.strip = False
+        self.upx = False
+
+        # .app bundle identifier for Code Signing
+        self.bundle_identifier = kws.get('bundle_identifier')
+        if not self.bundle_identifier:
+            # Fallback to appname.
+            self.bundle_identifier = self.appname
+
+        self.info_plist = kws.get('info_plist', None)
+
+        for arg in args:
+            if isinstance(arg, EXE):
+                self.toc.append((os.path.basename(arg.name), arg.name, arg.typ))
+                self.toc.extend(arg.dependencies)
+                self.strip = arg.strip
+                self.upx = arg.upx
+            elif isinstance(arg, TOC):
+                self.toc.extend(arg)
+                # TOC doesn't have a strip or upx attribute, so there is no way for us to
+                # tell which cache we should draw from.
+            elif isinstance(arg, COLLECT):
+                self.toc.extend(arg.toc)
+                self.strip = arg.strip_binaries
+                self.upx = arg.upx_binaries
+            else:
+                logger.info("unsupported entry %s", arg.__class__.__name__)
+        # Now, find values for app filepath (name), app name (appname), and name
+        # of the actual executable (exename) from the first EXECUTABLE item in
+        # toc, which might have come from a COLLECT too (not from an EXE).
+        for inm, name, typ in self.toc:
+            if typ == "EXECUTABLE":
+                self.exename = name
+                if self.name is None:
+                    self.appname = "Mac%s" % (os.path.splitext(inm)[0],)
+                    self.name = os.path.join(CONF['specpath'], self.appname + ".app")
+                else:
+                    self.name = os.path.join(CONF['specpath'], self.name)
+                break
+        self.__postinit__()
+
+    GUTS = (('toc', _check_guts_eq),  # additional check below
+            )
+
+    def check_guts(self, last_build):
+        # BUNDLE always needs to be executed, since it will clean the output
+        # directory anyway to make sure there is no existing cruft accumulating
+        return 1
+
+    def assemble(self):
+        if _check_path_overlap(self.name) and os.path.isdir(self.name):
+            _rmtree(self.name)
+        logger.info("Building BUNDLE %s", os.path.basename(self.out))
+
+        # Create a minimal Mac bundle structure
+        os.makedirs(os.path.join(self.name, "Contents", "MacOS"))
+        os.makedirs(os.path.join(self.name, "Contents", "Resources"))
+        os.makedirs(os.path.join(self.name, "Contents", "Frameworks"))
+
+        # Copy icns icon to Resources directory.
+        if os.path.exists(self.icon):
+            shutil.copy(self.icon, os.path.join(self.name, 'Contents', 'Resources'))
+        else:
+            logger.warn("icon not found %s" % self.icon)
+
+        # Key/values for a minimal Info.plist file
+        info_plist_dict = {"CFBundleDisplayName": self.appname,
+                           "CFBundleName": self.appname,
+
+                           # Required by 'codesign' utility.
+                           # The value for CFBundleIdentifier is used as the default unique
+                           # name of your program for Code Signing purposes.
+                           # It even identifies the APP for access to restricted OS X areas
+                           # like Keychain.
+                           #
+                           # The identifier used for signing must be globally unique. The usal
+                           # form for this identifier is a hierarchical name in reverse DNS
+                           # notation, starting with the toplevel domain, followed by the
+                           # company name, followed by the department within the company, and
+                           # ending with the product name. Usually in the form:
+                           #   com.mycompany.department.appname
+                           # Cli option --osx-bundle-identifier sets this value.
+                           "CFBundleIdentifier": self.bundle_identifier,
+
+                           # Fix for #156 - 'MacOS' must be in the name - not sure why
+                           "CFBundleExecutable": 'MacOS/%s' % os.path.basename(self.exename),
+                           "CFBundleIconFile": os.path.basename(self.icon),
+                           "CFBundleInfoDictionaryVersion": "6.0",
+                           "CFBundlePackageType": "APPL",
+                           "CFBundleShortVersionString": self.version,
+
+                           # Setting this to 1 will cause Mac OS X *not* to show
+                           # a dock icon for the PyInstaller process which
+                           # decompresses the real executable's contents. As a
+                           # side effect, the main application doesn't get one
+                           # as well, but at startup time the loader will take
+                           # care of transforming the process type.
+                           "LSBackgroundOnly": "0",
+
+                           }
+
+        # Merge info_plist settings from spec file
+        if isinstance(self.info_plist, dict) and self.info_plist:
+            info_plist_dict = dict(info_plist_dict.items() + self.info_plist.items())
+
+        info_plist = """<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple Computer//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>"""
+        for k, v in info_plist_dict.items():
+            info_plist += "<key>%s</key>\n<string>%s</string>\n" % (k, v)
+        info_plist += """</dict>
+</plist>"""
+        f = open(os.path.join(self.name, "Contents", "Info.plist"), "w")
+        f.write(info_plist)
+        f.close()
+
+        toc = add_suffix_to_extensions(self.toc)
+        for inm, fnm, typ in toc:
+            # Copy files from cache. This ensures that are used files with relative
+            # paths to dynamic library dependencies (@executable_path)
+            if typ in ('EXTENSION', 'BINARY'):
+                fnm = checkCache(fnm, strip=self.strip, upx=self.upx, dist_nm=inm)
+            tofnm = os.path.join(self.name, "Contents", "MacOS", inm)
+            todir = os.path.dirname(tofnm)
+            if not os.path.exists(todir):
+                os.makedirs(todir)
+            shutil.copy2(fnm, tofnm)
+
+        logger.info('moving BUNDLE data files to Resource directory')
+
+        ## For some hooks move resource to ./Contents/Resources dir.
+        # PyQt4/PyQt5 hooks: On Mac Qt requires resources 'qt_menu.nib'.
+        # It is moved from MacOS directory to Resources.
+        qt_menu_dir = os.path.join(self.name, 'Contents', 'MacOS', 'qt_menu.nib')
+        qt_menu_dest = os.path.join(self.name, 'Contents', 'Resources', 'qt_menu.nib')
+        if os.path.exists(qt_menu_dir):
+            shutil.move(qt_menu_dir, qt_menu_dest)
+
+        # Mac OS X Code Signing does not work when .app bundle contains
+        # data files in dir ./Contents/MacOS.
+        #
+        # Move all directories from ./MacOS/ to ./Resources and create symlinks
+        # in ./MacOS.
+        bin_dir =os.path.join(self.name, 'Contents', 'MacOS')
+        res_dir =os.path.join(self.name, 'Contents', 'Resources')
+        # Qt plugin directories does not contain data files.
+        ignore_dirs = set(['qt4_plugins', 'qt5_plugins'])
+        dirs = os.listdir(bin_dir)
+        for d in dirs:
+            abs_d = os.path.join(bin_dir, d)
+            res_d = os.path.join(res_dir, d)
+            if os.path.isdir(abs_d) and d not in ignore_dirs:
+                shutil.move(abs_d, res_d)
+                os.symlink(os.path.relpath(res_d, os.path.dirname(abs_d)), abs_d)
+
+        return 1
+
+
+class MERGE(object):
+    """
+    Merge repeated dependencies from other executables into the first
+    execuable. Data and binary files are then present only once and some
+    disk space is thus reduced.
+    """
+    def __init__(self, *args):
+        """
+        Repeated dependencies are then present only once in the first
+        executable in the 'args' list. Other executables depend on the
+        first one. Other executables have to extract necessary files
+        from the first executable.
+
+        args  dependencies in a list of (Analysis, id, filename) tuples.
+              Replace id with the correct filename.
+        """
+        # The first Analysis object with all dependencies.
+        # Any item from the first executable cannot be removed.
+        self._main = None
+
+        self._dependencies = {}
+
+        self._id_to_path = {}
+        for _, i, p in args:
+            self._id_to_path[i] = p
+
+        # Get the longest common path
+        self._common_prefix = os.path.dirname(os.path.commonprefix([os.path.abspath(a.scripts[-1][1]) for a, _, _ in args]))
+        if self._common_prefix[-1] != os.sep:
+            self._common_prefix += os.sep
+        logger.info("Common prefix: %s", self._common_prefix)
+
+        self._merge_dependencies(args)
+
+    def _merge_dependencies(self, args):
+        """
+        Filter shared dependencies to be only in first executable.
+        """
+        for analysis, _, _ in args:
+            path = os.path.abspath(analysis.scripts[-1][1]).replace(self._common_prefix, "", 1)
+            path = os.path.splitext(path)[0]
+            if path in self._id_to_path:
+                path = self._id_to_path[path]
+            self._set_dependencies(analysis, path)
+
+    def _set_dependencies(self, analysis, path):
+        """
+        Synchronize the Analysis result with the needed dependencies.
+        """
+        for toc in (analysis.binaries, analysis.datas):
+            for i, tpl in enumerate(toc):
+                if not tpl[1] in self._dependencies:
+                    logger.debug("Adding dependency %s located in %s" % (tpl[1], path))
+                    self._dependencies[tpl[1]] = path
+                else:
+                    dep_path = self._get_relative_path(path, self._dependencies[tpl[1]])
+                    logger.debug("Referencing %s to be a dependecy for %s, located in %s" % (tpl[1], path, dep_path))
+                    analysis.dependencies.append((":".join((dep_path, tpl[0])), tpl[1], "DEPENDENCY"))
+                    toc[i] = (None, None, None)
+            # Clean the list
+            toc[:] = [tpl for tpl in toc if tpl != (None, None, None)]
+
+    # TODO move this function to PyInstaller.compat module (probably improve
+    #      function compat.relpath()
+    # TODO use os.path.relpath instead
+    def _get_relative_path(self, startpath, topath):
+        start = startpath.split(os.sep)[:-1]
+        start = ['..'] * len(start)
+        if start:
+            start.append(topath)
+            return os.sep.join(start)
+        else:
+            return topath
+
+
+UNCOMPRESSED = 0
+COMPRESSED = 1
+
+
+class Tree(Target, TOC):
+    """
+    This class is a way of creating a TOC (Table of Contents) that describes
+    some or all of the files within a directory.
+    """
+    def __init__(self, root=None, prefix=None, excludes=None):
+        """
+        root
+                The root of the tree (on the build system).
+        prefix
+                Optional prefix to the names of the target system.
+        excludes
+                A list of names to exclude. Two forms are allowed:
+
+                    name
+                        Files with this basename will be excluded (do not
+                        include the path).
+                    *.ext
+                        Any file with the given extension will be excluded.
+        """
+        Target.__init__(self)
+        TOC.__init__(self)
+        self.root = root
+        self.prefix = prefix
+        self.excludes = excludes
+        if excludes is None:
+            self.excludes = []
+        self.__postinit__()
+
+    GUTS = (('root', _check_guts_eq),
+            ('prefix', _check_guts_eq),
+            ('excludes', _check_guts_eq),
+            ('toc', None),
+            )
+
+    def check_guts(self, last_build):
+        data = Target.get_guts(self, last_build)
+        if not data:
+            return True
+        stack = [data[0]]  # root
+        toc = data[3]  # toc
+        while stack:
+            d = stack.pop()
+            if misc.mtime(d) > last_build:
+                logger.info("Building %s because directory %s changed",
+                            self.outnm, d)
+                return True
+            for nm in os.listdir(d):
+                path = os.path.join(d, nm)
+                if os.path.isdir(path):
+                    stack.append(path)
+        self.data = toc
+        return False
+
+    def assemble(self):
+        logger.info("Building Tree %s", os.path.basename(self.out))
+        stack = [(self.root, self.prefix)]
+        excludes = {}
+        xexcludes = {}
+        for nm in self.excludes:
+            if nm[0] == '*':
+                xexcludes[nm[1:]] = 1
+            else:
+                excludes[nm] = 1
+        rslt = []
+        while stack:
+            dir, prefix = stack.pop()
+            for fnm in os.listdir(dir):
+                if excludes.get(fnm, 0) == 0:
+                    ext = os.path.splitext(fnm)[1]
+                    if xexcludes.get(ext, 0) == 0:
+                        fullfnm = os.path.join(dir, fnm)
+                        rfnm = prefix and os.path.join(prefix, fnm) or fnm
+                        if os.path.isdir(fullfnm):
+                            stack.append((fullfnm, rfnm))
+                        else:
+                            rslt.append((rfnm, fullfnm, 'DATA'))
+        self.data = rslt
+        try:
+            oldstuff = load_py_data_struct(self.out)
+        except:
+            oldstuff = None
+        newstuff = (self.root, self.prefix, self.excludes, self.data)
+        if oldstuff != newstuff:
+            save_py_data_struct(self.out, newstuff)
+            return 1
+        logger.info("%s no change!", self.out)
+        return 0
+
+
+_MISSING_BOOTLOADER_ERRORMSG = """
+Fatal error: PyInstaller does not include a pre-compiled bootloader for your
+platform. See <http://pythonhosted.org/PyInstaller/#building-the-bootloader>
+for more details and instructions how to build the bootloader.
+"""
