+++ PyInstaller/depend/analysis.py
@@ -37,12 +37,11 @@ import logging
 import os
 import re
 
-from .. import compat as compat
+from PyInstaller.building.datastruct import TOC
 from ..utils.misc import load_py_data_struct, get_code_object
 from ..lib.modulegraph.modulegraph import ModuleGraph
 from ..compat import importlib_load_source, is_py2, PY3_BASE_MODULES
 from .. import HOMEPATH
-
 
 logger = logging.getLogger(__name__)
 
@@ -332,95 +331,6 @@ class PyiModuleGraph(ModuleGraph):
                     rthooks_nodes.append(self.run_script(path))
 
         return rthooks_nodes
-
-
-# TODO Simplify the representation and use directly Modulegraph objects.
-class TOC(compat.UserList):
-    """
-    TOC (Table of Contents) class is a list of tuples of the form (name, path, tytecode).
-
-    typecode    name                   path                        description
-    --------------------------------------------------------------------------------------
-    EXTENSION   Python internal name.  Full path name in build.    Extension module.
-    PYSOURCE    Python internal name.  Full path name in build.    Script.
-    PYMODULE    Python internal name.  Full path name in build.    Pure Python module (including __init__ modules).
-    PYZ         Runtime name.          Full path name in build.    A .pyz archive (ZlibArchive data structure).
-    PKG         Runtime name.          Full path name in build.    A .pkg archive (Carchive data structure).
-    BINARY      Runtime name.          Full path name in build.    Shared library.
-    DATA        Runtime name.          Full path name in build.    Arbitrary files.
-    OPTION      The option.            Unused.                     Python runtime option (frozen into executable).
-
-    A TOC contains various types of files. A TOC contains no duplicates and preserves order.
-    PyInstaller uses TOC data type to collect necessary files bundle them into an executable.
-    """
-    def __init__(self, initlist=None):
-        compat.UserList.__init__(self)
-        self.filenames = set()
-        if initlist:
-            for entry in initlist:
-                self.append(entry)
-
-    def _normentry(self, entry):
-        if not isinstance(entry, tuple):
-            logger.info("TOC found a %s, not a tuple", entry)
-            raise TypeError("Expected tuple, not %s." % type(entry).__name__)
-        name, path, typecode = entry
-        if typecode in ["BINARY", "DATA"]:
-            # Normalize the case for binary files and data files only (to avoid duplicates
-            # for different cases under Windows). We can't do that for
-            # Python files because the import semantic (even at runtime)
-            # depends on the case.
-            name = os.path.normcase(name)
-        return (name, path, typecode)
-
-    def append(self, entry):
-        name, path, typecode = self._normentry(entry)
-        if name not in self.filenames:
-            self.data.append((name, path, typecode))
-            self.filenames.add(name)
-
-    def insert(self, pos, entry):
-        name, path, typecode = self._normentry(entry)
-        if name not in self.filenames:
-            self.data.insert(pos, (name, path, typecode))
-            self.filenames.add(name)
-
-    def __add__(self, other):
-        result = TOC(self)
-        result.extend(other)
-        return result
-
-    def __radd__(self, other):
-        result = TOC(other)
-        result.extend(self)
-        return result
-
-    def extend(self, other):
-        for entry in other:
-            self.append(entry)
-
-    def __sub__(self, other):
-        other = TOC(other)
-        filenames = self.filenames - other.filenames
-        result = TOC()
-        for name, path, typecode in self:
-            if name in filenames:
-                result.data.append((name, path, typecode))
-        return result
-
-    def __rsub__(self, other):
-        result = TOC(other)
-        return result.__sub__(self)
-
-    def intersect(self, other):
-        other = TOC(other)
-        filenames = self.filenames.intersection(other.filenames)
-        result = TOC()
-        for name, path, typecode in other:
-            if name in filenames:
-                result.data.append((name, path, typecode))
-        return result
-
 
 
 # TODO This class should raise exceptions on external callers attempting to
