+++ PyInstaller/utils/cliutils/archive_viewer.py
