+++ PyInstaller/utils/cliutils/grab_version.py
