+++ PyInstaller/utils/cliutils/pprint_toc.py
