+++ PyInstaller/utils/cliutils/set_version.py
