+++ utils/make_comserver.py
@@ -20,5 +20,5 @@ sys.path.insert(0, pyi_home)
 
 
 if __name__ == '__main__':
-    from PyInstaller.cliutils.make_comserver import run
+    from PyInstaller.utils.cliutils import run
     run()
